Reduce react refresh lint warnings
