+++ src/components/auth/ProtectedRoute.tsx
@@ -1,5 +1,5 @@
 import { Navigate, Outlet, useLocation } from 'react-router-dom';
-import { useAuth } from '@/contexts/AuthContext';
+import { useAuth } from '@/contexts/auth-context';
 
 export function ProtectedRoute() {
   const { isAuthenticated, loading } = useAuth();
