+++ src/components/portal/TechnicianManagementPanel.tsx
@@ -34,7 +34,7 @@ import {
 } from '@/components/ui/select';
 import { Skeleton } from '@/components/ui/skeleton';
 import { Textarea } from '@/components/ui/textarea';
-import { useAuth } from '@/contexts/AuthContext';
+import { useAuth } from '@/contexts/auth-context';
 import {
   fetchMyTechnicianGrants,
   fetchTechnicianManagementContext,
