+++ src/components/ui/sidebar.tsx
@@ -633,5 +633,4 @@ export {
   SidebarRail,
   SidebarSeparator,
   SidebarTrigger,
-  useSidebar,
 };
